Extract operation-related code to a new module
