+++ dhall/src/semantics/nze/nir.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 use std::rc::Rc;
 
+use crate::operations::OpKind;
 use crate::semantics::nze::lazy;
 use crate::semantics::{
     apply_any, normalize_hir, normalize_one_layer, squash_textlit, Binder,
@@ -8,7 +9,7 @@ use crate::semantics::{
 };
 use crate::syntax::{
     BinOp, Builtin, Const, Expr, ExprKind, InterpolatedTextContents, Label,
-    NumKind, OpKind, Span,
+    NumKind, Span,
 };
 use crate::ToExprOptions;
